+++ src/PSRule/Pipeline/GetRuleHelpPipelineBuilder.cs
@@ -10,10 +10,6 @@ namespace PSRule.Pipeline
         private readonly PipelineLogger _Logger;
 
         private RuleSource[] _Source;
-        private bool _LogError;
-        private bool _LogWarning;
-        private bool _LogVerbose;
-        private bool _LogInformation;
 
         internal GetRuleHelpPipelineBuilder()
         {
@@ -41,29 +37,24 @@ namespace PSRule.Pipeline
                 _Option.Baseline.Exclude = option.Baseline.Exclude;
             }
 
+            _Logger.Configure(_Option);
             return this;
         }
 
         public void UseCommandRuntime(ICommandRuntime2 commandRuntime)
         {
-            _Logger.OnWriteVerbose = commandRuntime.WriteVerbose;
-            _Logger.OnWriteWarning = commandRuntime.WriteWarning;
-            _Logger.OnWriteError = commandRuntime.WriteError;
-            _Logger.OnWriteInformation = commandRuntime.WriteInformation;
+            _Logger.UseCommandRuntime(commandRuntime);
         }
 
-        public void UseLoggingPreferences(ActionPreference error, ActionPreference warning, ActionPreference verbose, ActionPreference information)
+        public void UseExecutionContext(EngineIntrinsics executionContext)
         {
-            _LogError = (error != ActionPreference.Ignore);
-            _LogWarning = (warning != ActionPreference.Ignore);
-            _LogVerbose = !(verbose == ActionPreference.Ignore || verbose == ActionPreference.SilentlyContinue);
-            _LogInformation = !(information == ActionPreference.Ignore || information == ActionPreference.SilentlyContinue);
+            _Logger.UseExecutionContext(executionContext);
         }
 
         public GetRuleHelpPipeline Build()
         {
             var filter = new RuleFilter(ruleName: _Option.Baseline.RuleName, tag: null, exclude: _Option.Baseline.Exclude, wildcardMatch: true);
-            var context = PipelineContext.New(logger: _Logger, option: _Option, bindTargetName: null, bindTargetType: null, logError: _LogError, logWarning: _LogWarning, logVerbose: _LogVerbose, logInformation: _LogInformation);
+            var context = PipelineContext.New(logger: _Logger, option: _Option, bindTargetName: null, bindTargetType: null);
             return new GetRuleHelpPipeline(option: _Option, source: _Source, filter: filter, context: context);
         }
     }