+++ tests/PSRule.Tests/FromFileWithLogging.Rule.ps1
@@ -3,6 +3,7 @@
 #
 
 # Logging in main script
+Write-Debug -Message 'Script debug message';
 Write-Information -MessageData 'Script information message';
 Write-Verbose -Message 'Script verbose message';
 Write-Warning -Message 'Script warning message';
@@ -22,7 +23,22 @@ Rule 'WithVerbose' {
     $True;
 }
 
+Rule 'WithVerbose2' {
+    Write-Verbose -Message 'Rule verbose message 2';
+    $True;
+}
+
 Rule 'WithInformation' {
     Write-Information -MessageData 'Rule information message';
+    $True;
+}
+
+Rule 'WithDebug' {
+    Write-Debug -Message 'Rule debug message';
+    $True;
+}
+
+Rule 'WithDebug2' {
+    Write-Debug -Message 'Rule debug message 2';
     $True;
 }